==> Python/01_parameters.ipynb <==
import os
import itertools

import numpy as np
import pandas as pd

# Generate parameter grid for the different analyses #####################################

# Expand grid function from dictionary of lists
def expand_grid(dict):
    rows = itertools.product(*dict.values())
    return pd.DataFrame.from_records(rows, columns = dict.keys())

# Load data source for grid
Y_df = pd.read_csv(os.path.join("data", "pheno_original.csv"))
parts = pd.read_csv(os.path.join("data", "partitions.csv"))

# Traits list
traits = Y_df.columns[[3, 9, 10, 12]].to_list()

# 1. top-markers-10k -------------------------------------------------------------------------
model_type = "top-markers-10k"

# 1.1. AroAdm ============================================================================
out_name = f"AroAdm_{model_type}.csv"

folds = ["AroAdm"]

grid_dict = {
    "model_type": [model_type],
    "architecture": ["MLP", "CNN"],
    "m_set": ["SNP", "all"],
    "traits": traits,
    "part": folds,
    "input_type": ["single-input"],
    "output_type": ["single-output"]
}

grid = expand_grid(grid_dict)
print(grid)

grid.to_csv(os.path.join("parameters", out_name), index = False)

# 1.2. 10-fold-cv ========================================================================
out_name = f"10-fold-cv_{model_type}.csv"

folds = parts.columns.drop(["Accession", "AroAdm"]).to_list()

grid_dict = {
    "model_type": [model_type],
    "architecture": ["MLP", "CNN"],
    "m_set": ["SNP", "all"],
    "traits": traits,
    "part": folds,
    "input_type": ["single-input"],
    "output_type": ["single-output"]
}

grid = expand_grid(grid_dict)
print(grid)

grid.to_csv(os.path.join("parameters", out_name), index = False)

# 2. kerPC -------------------------------------------------------------------------------
model_type = "kerPC"

# 2.1. AroAdm ============================================================================
out_name = f"AroAdm_{model_type}.csv"

folds = ["AroAdm"]

grid_dict = {
    "model_type": [model_type],
    "architecture": ["MLP"],
    "m_set": ["SNP", "all"],
    "traits": traits,
    "part": folds,
    "input_type": ["single-input", "multi-input"],
    "output_type": ["single-output"]
}

grid = expand_grid(grid_dict)

# No multi-input with SNP kernel case
grid = grid[~ ((grid.input_type == "multi-input") & (grid.m_set == "SNP"))]
print(grid)

grid.to_csv(os.path.join("parameters", out_name), index = False)

# 2.2. 10-fold-cv ========================================================================
out_name = f"10-fold-cv_{model_type}.csv"

folds = parts.columns.drop(["Accession", "AroAdm"]).to_list()

grid_dict = {
    "model_type": [model_type],
    "architecture": ["MLP"],
    "m_set": ["SNP", "all"],
    "traits": traits,
    "part": folds,
    "input_type": ["single-input", "multi-input"],
    "output_type": ["single-output"]
}

grid = expand_grid(grid_dict)

# No multi-input with SNP kernel case
grid = grid[~ ((grid.input_type == "multi-input") & (grid.m_set == "SNP"))]
print(grid)

grid.to_csv(os.path.join("parameters", out_name), index = False)

# 3. multi-output ------------------------------------------------------------------------

target_dict = {
    "culm.diameter.1st.internode": "binary",
    "leaf.senescence": "binary",
    "grain.weight": "continuous",
    "time.to.flowering.from.sowing": "continuous"
}

# Lists of traits to predict jointly
trait_lists = []
# Combinations of 2 with same variable type (binary or continuous)
for comb in itertools.combinations(traits, 2):
    target_type_list = []
    for trait in comb:
        target_type_list.append(target_dict[trait])
    if all(x == target_type_list[0] for x in target_type_list):
        trait_lists.append(list(comb))
# All traits at the same time
trait_lists.append(traits)
        
print(trait_lists)

# 3.1 top-markers-10k ========================================================================
model_type = "top-markers-10k"
output_type = "multi-output"
input_type = "multi-input"

# AroAdm #################################################################################
out_name = f"AroAdm_{model_type}_{input_type}_{output_type}.csv"

folds = ["AroAdm"]

grid_dict = {
    "model_type": [model_type],
    "architecture": ["MLP"],
    "m_set": ["SNP", "all"],
    "traits": trait_lists,
    "part": folds,
    "input_type": [input_type],
    "output_type": [output_type]
}

grid = expand_grid(grid_dict)
print(grid)

grid.to_csv(os.path.join("parameters", out_name), index = False)

# 10-fold-cv #############################################################################
out_name = f"10-fold-cv_{model_type}_{input_type}_{output_type}.csv"

folds = parts.columns.drop(["Accession", "AroAdm"]).to_list()

grid_dict = {
    "model_type": [model_type],
    "architecture": ["MLP"],
    "m_set": ["SNP", "all"],
    "traits": trait_lists,
    "part": folds,
    "input_type": [input_type],
    "output_type": [output_type]
}

grid = expand_grid(grid_dict)
print(grid)

grid.to_csv(os.path.join("parameters", out_name), index = False)

# 3.2 kerPC ==============================================================================
model_type = "kerPC"
output_type = "multi-output"

# AroAdm #################################################################################
out_name = f"AroAdm_{model_type}_{output_type}.csv"

folds = ["AroAdm"]

grid_dict = {
    "model_type": [model_type],
    "architecture": ["MLP"],
    "m_set": ["SNP", "all"],
    "traits": trait_lists,
    "part": folds,
    "input_type": ["single-input", "multi-input"],
    "output_type": [output_type]
}

grid = expand_grid(grid_dict)

# No multi-input with SNP kernel case
grid = grid[~ ((grid.input_type == "multi-input") & (grid.m_set == "SNP"))]
print(grid)

grid.to_csv(os.path.join("parameters", out_name), index = False)

# 10-fold-cv #############################################################################
out_name = f"10-fold-cv_{model_type}_{output_type}.csv"

folds = parts.columns.drop(["Accession", "AroAdm"]).to_list()

grid_dict = {
    "model_type": [model_type],
    "architecture": ["MLP"],
    "m_set": ["SNP", "all"],
    "traits": trait_lists,
    "part": folds,
    "input_type": ["single-input", "multi-input"],
    "output_type": [output_type]
}

grid = expand_grid(grid_dict)

# No multi-input with SNP kernel case
grid = grid[~ ((grid.input_type == "multi-input") & (grid.m_set == "SNP"))]
print(grid)

grid.to_csv(os.path.join("parameters", out_name), index = False)
