==> Python/05_get_results.ipynb <==
import os
import ast

import numpy as np
import pandas as pd

from matplotlib import pyplot as plt
plt.style.use("seaborn-whitegrid")
import seaborn as sns
sns.set_style("whitegrid")

# 1. Obtain parameters for the different analyses ----------------------------------------

par_file_list = [
    "10-fold-cv_kerPC.csv",
    "10-fold-cv_kerPC_multi-output.csv",
    "10-fold-cv_top-markers-10k.csv",
    "10-fold-cv_top-markers-10k_multi-input_multi-output.csv",
    "AroAdm_kerPC.csv",
    "AroAdm_kerPC_multi-output.csv",
    "AroAdm_top-markers-10k.csv",
    "AroAdm_top-markers-10k_multi-input_multi-output.csv"
]

# par_file_list = [
#     "10-fold-cv_kerPC.csv"
# ]

# Final dataframe
RES = pd.DataFrame()

# For each parameters (experiment) file
for par_file in par_file_list:
    
    # Load grid
    par_grid = pd.read_csv(os.path.join("parameters", par_file))

    # For each row of grid (each analysis)
    for row_number in range(par_grid.shape[0]):
        
        # Extract parameters =====================================================================

        # Which row of the grid
        pars = par_grid.iloc[row_number]
        # print(pars, "\n")

        # Model type: top-markers-10k or kerPC
        model_type = pars["model_type"]

        # Common pars
        architecture = pars["architecture"]
        m_set = pars["m_set"]
        traits = pars["traits"]
        part = pars["part"]
        input_type = pars["input_type"]
        output_type = pars["output_type"]

        # Output folder
        out_folder = os.path.join("models", f"{model_type}_{architecture}_{m_set}_{input_type}", traits, part)
        if not os.path.isdir(out_folder):
            os.makedirs(out_folder)
        print(out_folder)

        # Save traits as list
        # If multi-output, change literal string to list
        if output_type == "multi-output":
            traits = ast.literal_eval(traits)
        else:
            traits = [traits]
        # print(traits)

        # Target variable type
        # Save as list, just like traits
        target_dict = {
            "culm.diameter.1st.internode": "binary",
            "leaf.senescence": "binary",
            "grain.weight": "continuous",
            "time.to.flowering.from.sowing": "continuous"
        }

        target_types = []
        for trait_i in traits:
            if trait_i not in target_dict.keys():
                print("Define variable type of the target in the dictionary")
                exit(1)
            target_types.append(target_dict[trait_i])

        # print(target_types)
        
        # 2. Get prediction results ------------------------------------------------------

        # Which partition was used for Hpar optimization
        # which_hps = part
        which_hps = "AroAdm"
        # print(which_hps)

        # Path of prediction results
        pred_path = os.path.join(out_folder, f"results_hps_{which_hps}")
        # print(pred_path)

        # Prediction in case we wanted it
        # for i in range(len(traits)):
        #     prediction_path = os.path.join(pred_path, f"prediction_{traits[i]}.csv")
        #     prediction = pd.read_csv(prediction_path)
        
        # Results file
        res_path = os.path.join(pred_path, "results.csv")
        res = pd.read_csv(res_path, index_col = 0)
        res = res["mean"].to_frame().T.reset_index(drop = True)
        
        # Experiment parameters
        pars_tmp = pars.to_frame().T.reset_index(drop = True)
        
        # Join experiment pars and results
        res_tmp = pd.concat([pars_tmp, res], axis = 1)
        
        # Row bind to final dataframe
        RES = pd.concat([RES, res_tmp], ignore_index = True, sort = False)


RES

# Save results

final_path = os.path.join("final_results")
if not os.path.isdir(final_path):
    os.mkdir(final_path)

RES.to_csv(os.path.join(final_path, "final_results.csv"), index = False)
